Break up text in info section
Why: Design requires that box and text of info section are at the same width. This means that the text would break/wrap

To solve this, max-width is added to info such that the text would occupy more or less the same width with box and break-up to next line.